Move from domain.usecases to usecases
