+++ src/main/java/br/com/fiap/soat/techChallenge/usecases/model/ItemDoComandoDeNovoPedido.java
@@ -1,4 +1,4 @@
-package br.com.fiap.soat.techChallenge.domain.usecases.model;
+package br.com.fiap.soat.techChallenge.usecases.model;
 
 import java.util.UUID;
